Add container offset to scrollbar length
If the `#container` element is set in order to have some fixed
content before the iterated elements, the calculated length of the
scroll bar was just the sum of the lengths of the iterated elements,
without the size of the fixed part. This means that the scroll bar is
too short, and even though we fully scrolled down, more content is
available below the last element.

Calculate an additional scrollbarLength property that includes the
fixed content offset and use it for the scrollbar.